re-order files to support docker-compose debugging
