+++ .env
@@ -11,10 +11,4 @@ CONTAINER_PROJECT_NAME=kiss-frontend
 
 # specific settings for the front end / vite
 VITE_API_REFRESH_INTERVAL_MS=30000
-VITE_API_BASE_URI=http://localhost/api/nieuws
-
-##################################################
-# RabbitMQ
-##################################################
-RABBITMQ_USERNAME=common-gateway
-RABBITMQ_PASSWORD=changeme!
+BFF_SSL_PORT=7231
