--- DockerfileDev
@@ -1,12 +1,0 @@
-FROM node:14 AS builder
-WORKDIR /app
-COPY . .
-
-RUN npm ci
-RUN npm run build --if-present
-RUN npm test
-
-FROM nginx:alpine
-EXPOSE 80
-COPY docker/nginx/default.conf.template /etc/nginx/templates/default.conf.template
-COPY --from=builder /app/dist /usr/share/nginx/html
