+++ Kiss.Bff/Properties/launchSettings.json
@@ -8,7 +8,7 @@
         "ASPNETCORE_ENVIRONMENT": "Development"
       },
       "dotnetRunMessages": true,
-      "applicationUrl": "https://localhost:7230;http://localhost:5053"
+      "applicationUrl": "https://localhost:7081;http://localhost:5072"
     },
     "IIS Express": {
       "commandName": "IISExpress",
@@ -31,8 +31,8 @@
     "windowsAuthentication": false,
     "anonymousAuthentication": true,
     "iisExpress": {
-      "applicationUrl": "http://localhost:26740",
-      "sslPort": 44349
+      "applicationUrl": "http://localhost:55714",
+      "sslPort": 44361
     }
   }
 }
