--- docker-compose.deploy.yaml
@@ -1,8 +1,0 @@
-version: "3.4"
-
-services:
-  frontend:
-    image: ${CONTAINER_REGISTRY_BASE}/${CONTAINER_PROJECT_NAME}:${APP_BUILD}
-    build:
-      context: .
-      dockerfile: Dockerfile
