--- docker/nginx/default.conf.template
@@ -1,14 +1,0 @@
-server {
-    root /usr/share/nginx/html;
-    location / {
-        try_files $uri $uri/ /index.html$is_args$args;
-    }
-
-    location = /index.html {
-        internal;
-        add_header "Cache-Control" "no-cache" ;
-    }
-    location = /env.js {
-        add_header "Cache-Control" "no-cache" ;
-    }
-}
