+++ launchSettings.json
@@ -1,0 +1,14 @@
+{
+  "profiles": {
+    "Docker Compose": {
+      "commandName": "DockerCompose",
+      "commandVersion": "1.0",
+      "composeLaunchAction": "LaunchBrowser",
+      "composeLaunchServiceName": "kiss.bff",
+      "composeLaunchUrl": "{Scheme}://localhost:{ServicePort}/weatherforecast",
+      "serviceActions": {
+        "kiss.bff": "StartDebugging"
+      }
+    }
+  }
+}
