--- user-component/default.yaml
@@ -1,43 +1,0 @@
-App\Entity\Group:
-    admin:
-        description: the admin group
-        id: 2809c0d7-fecf-41da-8dbf-3d09af0d025c
-        name: admin
-        organization: http://api/admin/organizations/a1c8e0b6-2f78-480d-a9fb-9792142f4761
-        scopes:
-            - '@admin_create'
-            - '@admin_read'
-            - '@admin_remove'
-            - '@admin_update'
-    anonymous:
-        description: the anonymous group
-        name: ANONYMOUS
-        organization: http://api/admin/organizations/a1c8e0b6-2f78-480d-a9fb-9792142f4761
-        scopes:
-            - '@admin_create'
-            - '@admin_read'
-            - '@admin_remove'
-            - '@admin_update'
-App\Entity\Scope:
-    admin_create:
-        code: POST.admin
-        name: Create admin
-    admin_read:
-        code: GET.admin
-        name: Read admin
-    admin_remove:
-        code: DELETE.admin
-        name: Remove admin
-    admin_update:
-        code: PUT.admin
-        name: Update admin
-App\Entity\User:
-    someUser:
-        id: 2ac800cc-b844-458d-87e1-b206fa80113a
-        organization: http://api/admin/organizations/a1c8e0b6-2f78-480d-a9fb-9792142f4761
-        #!ChangeMe!
-        password: '\$2y\$10\$j.1PMChAalEoRYgVd3xAN.6Ikk8p0MfiknhLGrp8bU./SjxXUa6pG'
-        person: http://api/admin/people/d2a24edf-6ced-40a2-8e7c-56803f4ce3f5
-        userGroups:
-            - '@admin'
-        username: test@gateway.local
